1.0.5 version

+++ app/src/main/java/com/skywavestudios/prefactor/Product.java
@@ -1,12 +1,14 @@
 package com.skywavestudios.prefactor;
 
+import java.math.BigInteger;
+
 /**
  * Created by Iman on 4/9/2017.
  */
 
 public class Product {
     public String Name;
-    public int Fee;
+    public long Fee;
     public float No;
     public String Scale;
 }
